Prettify

+++ contracts/staking/original/dependency/Pausable.sol
@@ -6,7 +6,7 @@ import "./Owned.sol";
 
 // https://docs.synthetix.io/contracts/Pausable
 contract Pausable is Owned {
-    uint public lastPauseTime;
+    uint256 public lastPauseTime;
     bool public paused;
 
     constructor() internal {
@@ -40,7 +40,10 @@ contract Pausable is Owned {
     event PauseChanged(bool isPaused);
 
     modifier notPaused {
-        require(!paused, "This action cannot be performed while the contract is paused");
+        require(
+            !paused,
+            "This action cannot be performed while the contract is paused"
+        );
         _;
     }
 }
